# src/openapi/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Texas Help 'Em API",
        "description": "Manage a players hand in a given game of Texas Hold 'Em",
        "version": "1.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "API Support",
            "email": "api@adrianroworth.co.uk"
        }
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
...
    "paths": {
        "/hands/": {
            "post": {
                "tags": [
                    "Hands"
                ],
                "summary": "Gets all valid hand types for the set of cards posted",
                "operationId": "getHands",
                "x-requests": {
                    "400": {
                        "requestBody": {
                            "content": {
                                "application/vnd.api+json": {
                                    "example": {
                                        "data": {
                                            "cards": [
                                                "20-s",
                                                "3-x",
                                                "3-h",
                                                "10-s",
                                                "9-p",
                                                "6-s",
                                                "4-c"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "description": "Array of 7 card IDs",
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": false,
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "object",
                                        "additionalProperties": false,
                                        "required": [
                                            "cards"
                                        ],
                                        "properties": {
                                            "cards": {
                                                "type": "array",
                                                "uniqueItems": true,
                                                "minItems": 7,
                                                "maxItems": 7,
                                                "items": {
                                                    "type": "string",
                                                    "pattern": "^([1-9]|1[0-2])-[s,h,d,c]$"
                                                }
                                            }
                                        }
                                    }
                                }
                            },
                            "example": {
                                "data": {
                                    "cards": [
                                        "1-s",
                                        "3-s",
                                        "3-h",
                                        "10-s",
                                        "9-d",
                                        "6-s",
                                        "4-c"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "required": [
                                                    "type",
                                                    "id",
                                                    "attributes"
                                                ],
                                                "additionalProperties": false,
                                                "properties": {
                                                    "type": {
                                                        "enum": [
                                                            "hands"
                                                        ]
                                                    },
                                                    "id": {
                                                        "type": "string",
                                                        "enum": [
                                                            "high-card",
                                                            "pair",
                                                            "two-pairs",
                                                            "three-of-a-kind",
                                                            "straight",
                                                            "flush",
                                                            "full-house",
                                                            "four-of-a-kind",
                                                            "straight-flush",
                                                            "royal-flush"
                                                        ]
                                                    },
                                                    "attributes": {
                                                        "type": "object",
                                                        "required": [
                                                            "cards",
                                                            "description",
                                                            "precedence",
                                                            "name"
                                                        ],
                                                        "additionalProperties": false,
                                                        "properties": {
                                                            "cards": {
                                                                "type": "object",
                                                                "additionalProperties": false,
                                                                "required": [
                                                                    "best",
                                                                    "all"
                                                                ],
                                                                "properties": {
                                                                    "best": {
                                                                        "type": "object",
                                                                        "required": [
                                                                            "cardCode",
                                                                            "cardValue",
                                                                            "cardSuit",
                                                                            "cardName",
                                                                            "card"
                                                                        ],
                                                                        "additionalProperties": false,
                                                                        "properties": {
                                                                            "cardCode": {
                                                                                "type": "string"
                                                                            },
                                                                            "cardValue": {
                                                                                "type": "integer"
                                                                            },
                                                                            "cardSuit": {
                                                                                "type": "string"
                                                                            },
                                                                            "cardName": {
                                                                                "type": "string"
                                                                            },
                                                                            "card": {
                                                                                "type": "string"
                                                                            }
                                                                        }
                                                                    },
                                                                    "all": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "object",
                                                                            "required": [
                                                                                "cardCode",
                                                                                "cardValue",
                                                                                "cardSuit",
                                                                                "cardName",
                                                                                "card"
                                                                            ],
                                                                            "additionalProperties": false,
                                                                            "properties": {
                                                                                "cardCode": {
                                                                                    "type": "string"
                                                                                },
                                                                                "cardValue": {
                                                                                    "type": "integer"
                                                                                },
                                                                                "cardSuit": {
                                                                                    "type": "string"
                                                                                },
                                                                                "cardName": {
                                                                                    "type": "string"
                                                                                },
                                                                                "card": {
                                                                                    "type": "string"
                                                                                }
                                                                            }
                                                                        }
                                                                    }
                                                                }
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "maxLength": 200
                                                            },
                                                            "precedence": {
                                                                "type": "number",
                                                                "minimum": 0,
                                                                "maximum": 9
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "maxLength": 20
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "hands",
                                            "id": "high-card",
                                            "attributes": {
                                                "cards": {
                                                    "best": {
                                                        "cardCode": "1-s",
                                                        "cardValue": 1,
                                                        "cardSuit": "spades",
                                                        "cardName": "ace",
                                                        "card": "ace of spades"
                                                    },
                                                    "all": [
                                                        {
                                                            "cardCode": "1-s",
                                                            "cardValue": 1,
                                                            "cardSuit": "spades",
                                                            "cardName": "ace",
                                                            "card": "ace of spades"
                                                        }
                                                    ]
                                                },
                                                "description": "High ace",
                                                "precedence": 0,
                                                "name": "High Card"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "There is an issue with the request",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "errors"
                                    ],
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "required": [
                                                    "status",
                                                    "title",
                                                    "detail"
                                                ],
                                                "properties": {
                                                    "status": {
                                                        "enum": [
                                                            400
                                                        ]
                                                    },
                                                    "title": {
                                                        "enum": [
                                                            "400 Bad Request"
                                                        ]
                                                    },
                                                    "detail": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "errors": [
                                        {
                                            "status": 400,
                                            "title": "400 Bad Request",
                                            "detail": "Request JSON is malformed"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The specified resource was not found",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "errors"
                                    ],
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "required": [
                                                    "status",
                                                    "title",
                                                    "detail"
                                                ],
                                                "properties": {
                                                    "status": {
                                                        "enum": [
                                                            404
                                                        ]
                                                    },
                                                    "title": {
                                                        "enum": [
                                                            "404 Not Found"
                                                        ]
                                                    },
                                                    "detail": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "errors": [
                                        {
                                            "status": 404,
                                            "title": "404 Not Found",
                                            "detail": "Resource /api/v1/url does not exist"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}